requirements.txt
asgiref==3.5.2
attrs==21.4.0
Django==3.0
djangorestframework==3.13.1
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
pytest==7.1.2
pytest-django==4.5.2
pytz==2022.1
sqlparse==0.4.2
tomli==2.0.1
certifi==2022.6.15
distlib==0.3.5
filelock==3.7.1
pipenv==2022.7.24
platformdirs==2.5.2
six==1.16.0
virtualenv==20.15.1
virtualenv-clone==0.5.7
asgiref==3.5.2
attrs==21.4.0
certifi==2022.6.15
charset-normalizer==2.1.0
distlib==0.3.5
Django==3.0
djangorestframework==3.13.1
filelock==3.7.1
idna==3.3
iniconfig==1.1.1
packaging==21.3
pipenv==2022.7.24
platformdirs==2.5.2
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
pytest==7.1.2
pytest-django==4.5.2
pytz==2022.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
urllib3==1.26.11
virtualenv==20.15.1
virtualenv-clone==0.5.7
asgiref==3.5.2
attrs==21.4.0
certifi==2022.6.15
charset-normalizer==2.1.0
distlib==0.3.5
Django==3.0
djangorestframework==3.13.1
filelock==3.7.1
idna==3.3
iniconfig==1.1.1
jsonfield==3.1.0
packaging==21.3
pipenv==2022.7.24
platformdirs==2.5.2
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
pytest==7.1.2
pytest-django==4.5.2
pytz==2022.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
urllib3==1.26.11
virtualenv==20.15.1
virtualenv-clone==0.5.7
